More refactoring of Gauss Newton and now have a proper test.

+++ CentralDifferenceDifferentiator.h
@@ -11,11 +11,11 @@ class CentralDifferencesDifferentiator {
       cubeSize(volume->cubeSize) {}
 
     virtual void xDerivative(VolumeT *dx) const {
-      for(size_t z = 0; z < cubeSize; z++) {
-        for(size_t y = 0; y < cubeSize; y++) {
-          for(size_t x = 0; x < cubeSize; x++) {
-            size_t xMinus1 = volume->wrapIndex(x - 1);
-            size_t xPlus1 = volume->wrapIndex(x + 1);
+      for(int z = 0; z < cubeSize; z++) {
+        for(int y = 0; y < cubeSize; y++) {
+          for(int x = 0; x < cubeSize; x++) {
+            int xMinus1 = volume->wrapIndex(x - 1);
+            int xPlus1 = volume->wrapIndex(x + 1);
 
             dx->at(z, y, x) = (
                 volume->at(z, y, xPlus1) -
@@ -27,12 +27,12 @@ class CentralDifferencesDifferentiator {
     }
 
     virtual void yDerivative(VolumeT *dy) const {
-      for(size_t z = 0; z < cubeSize; z++) {
-        for(size_t y = 0; y < cubeSize; y++) {
-          size_t yMinus1 = volume->wrapIndex(y - 1);
-          size_t yPlus1 = volume->wrapIndex(y + 1);
+      for(int z = 0; z < cubeSize; z++) {
+        for(int y = 0; y < cubeSize; y++) {
+          int yMinus1 = volume->wrapIndex(y - 1);
+          int yPlus1 = volume->wrapIndex(y + 1);
 
-          for(size_t x = 0; x < cubeSize; x++) {
+          for(int x = 0; x < cubeSize; x++) {
             dy->at(z, y, x) = (
                 volume->at(z, yPlus1, x) -
                 volume->at(z, yMinus1, x)
@@ -43,12 +43,12 @@ class CentralDifferencesDifferentiator {
     }
 
     virtual void zDerivative(VolumeT *dz) const {
-      for(size_t z = 0; z < cubeSize; z++) {
-        size_t zMinus1 = volume->wrapIndex(z - 1);
-        size_t zPlus1 = volume->wrapIndex(z + 1);
+      for(int z = 0; z < cubeSize; z++) {
+        int zMinus1 = volume->wrapIndex(z - 1);
+        int zPlus1 = volume->wrapIndex(z + 1);
 
-        for(size_t y = 0; y < cubeSize; y++) {
-          for(size_t x = 0; x < cubeSize; x++) {
+        for(int y = 0; y < cubeSize; y++) {
+          for(int x = 0; x < cubeSize; x++) {
             dz->at(z, y, x) = (
                 volume->at(zPlus1, y, x) -
                 volume->at(zMinus1, y, x)
@@ -60,7 +60,7 @@ class CentralDifferencesDifferentiator {
 
   protected:
     const VolumeT *volume;
-    const size_t cubeSize; 
+    const int cubeSize; 
 };
 
 #endif
